+++ .gitignore
@@ -7,6 +7,7 @@ yarn-error.log
 .ionide/
 .metals/
 .ultra.cache.json
+.vscode/
 tsbuildinfo/
 del-tags.sh
 incubation/
