+++ packages/system/src/Effect/index.ts
@@ -15,6 +15,7 @@ export * from "./bracketExit"
 export * from "./bracketFiber"
 export * from "./bracketOnError"
 export * from "./cached"
+export * from "./cachedInvalidate"
 export * from "./Canceler"
 export * from "./catchAll"
 export * from "./catchAllCause_"
